snapshot: move process code around some more
