+++ app/process/flags.go
@@ -27,7 +27,7 @@ func DefaultFlagSet() *flag.FlagSet {
 
 	fs := flagset.NewFlagSet("server")
 
-	fs.StringVar(&mode, "mode", "cli", "Valid options are: cli, lambda")
+	fs.StringVar(&mode, "mode", "cli", "Valid options are: cli, fsnotify, lambda")
 	fs.BoolVar(&verbose, "verbose", false, "Enabled verbose (debug) loggging.")
 
 	fs.StringVar(&config_source, "config-source", "defaults://", "A valid Go Cloud bucket URI where your go-iiif config file is located. Optionally, if 'defaults://' is specified then the default config bundled with this package will be used.")
