+++ app/process/process.go
@@ -15,7 +15,6 @@ package process
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
@@ -30,102 +29,10 @@ import (
 	"github.com/fsnotify/fsnotify"
 	iiifuri "github.com/go-iiif/go-iiif-uri"
 	iiifaws "github.com/go-iiif/go-iiif/v6/aws"
-	iiifcache "github.com/go-iiif/go-iiif/v6/cache"
-	iiifconfig "github.com/go-iiif/go-iiif/v6/config"
 	iiifdriver "github.com/go-iiif/go-iiif/v6/driver"
 	iiifprocess "github.com/go-iiif/go-iiif/v6/process"
 	"gocloud.dev/blob"
 )
-
-type ProcessResultsReport map[string]interface{}
-
-type ProcessOptions struct {
-	Config             *iiifconfig.Config
-	Driver             iiifdriver.Driver
-	Processor          iiifprocess.Processor
-	Instructions       iiifprocess.IIIFInstructionSet
-	Report             bool
-	ReportTemplate     string
-	ReportBucket       *blob.Bucket
-	GenerateReportHTML bool
-}
-
-func ProcessMany(ctx context.Context, opts *ProcessOptions, uris ...iiifuri.URI) error {
-
-	_, err := ProcessManyWithReport(ctx, opts, uris...)
-	return err
-}
-
-func ProcessManyWithReport(ctx context.Context, opts *ProcessOptions, uris ...iiifuri.URI) (*ProcessResultsReport, error) {
-
-	results := make(ProcessResultsReport)
-
-	wg := new(sync.WaitGroup)
-
-	for _, uri := range uris {
-
-		origin := uri.Origin()
-
-		logger := slog.Default()
-		logger = logger.With("origin", origin)
-
-		rsp, err := iiifprocess.ParallelProcessURIWithInstructionSet(opts.Config, opts.Driver, opts.Processor, opts.Instructions, uri)
-
-		if err != nil {
-			return &results, err
-		}
-
-		if opts.Report {
-
-			uri_opts := &url.Values{}
-			uri_opts.Set("format", "jpg") // this is made up (and not necessarily part of the instructions file)
-			uri_opts.Set("label", "x")    // this is made up (and not necessarily part of the instructions file)
-
-			target, err := uri.Target(uri_opts)
-
-			if err != nil {
-				logger.Error("Unable to generate target URL for report", "error", err)
-			} else {
-
-				report_name, err := iiifprocess.DeriveReportNameFromURI(ctx, uri, opts.ReportTemplate)
-
-				if err == nil {
-
-					var root string
-
-					switch uri.(type) {
-					case *iiifuri.IdSecretURI:
-						root = filepath.Dir(target)
-					default:
-						root = target
-					}
-
-					key := filepath.Join(root, report_name)
-					wg.Add(1)
-
-					go func() {
-
-						defer wg.Done()
-						err := report_processing(ctx, opts, key, rsp)
-
-						if err != nil {
-							logger.Error("Unable to write process report", "key", key, "error", err)
-						}
-					}()
-
-				} else {
-					logger.Error("Unable to generate report name", "error", err)
-				}
-			}
-		}
-
-		results[origin] = rsp
-	}
-
-	wg.Wait()
-
-	return &results, nil
-}
 
 func Run(ctx context.Context) error {
 	fs := DefaultFlagSet()
@@ -202,7 +109,7 @@ func RunWithOptions(ctx context.Context, opts *RunOptions) error {
 
 		for _, str_uri := range opts.Paths {
 
-			u, err := t.URIFunc(str_uri)
+			u, err := iiifuri.NewURI(ctx, str_uri)
 
 			if err != nil {
 				return fmt.Errorf("URI Func for '%s' failed: %w", str_uri, err)
@@ -266,7 +173,7 @@ func RunWithOptions(ctx context.Context, opts *RunOptions) error {
 						rel_path := strings.Replace(abs_path, root, "", 1)
 						rel_path = strings.TrimLeft(rel_path, "/")
 
-						u, err := t.URIFunc(rel_path)
+						u, err := iiifuri.NewURI(ctx, rel_path)
 
 						if err != nil {
 							logger.Warn("Failed to parse path", "rel path", rel_path, "abs path", abs_path, "error", err)
@@ -316,7 +223,7 @@ func RunWithOptions(ctx context.Context, opts *RunOptions) error {
 
 				s3_fname := filepath.Base(s3_key)
 
-				u, err := t.URIFunc(s3_fname)
+				u, err := iiifuri.NewURI(ctx, s3_fname)
 
 				if err != nil {
 					return err
@@ -338,75 +245,6 @@ func RunWithOptions(ctx context.Context, opts *RunOptions) error {
 
 	default:
 		return errors.New("Unsupported mode")
-	}
-
-	return nil
-}
-
-func report_processing(ctx context.Context, opts *ProcessOptions, key string, rsp map[string]interface{}) error {
-
-	rsp_body, err := json.Marshal(rsp)
-
-	if err != nil {
-		return fmt.Errorf("Failed to marshal processing report, %w", err)
-	}
-
-	if opts.ReportBucket != nil {
-
-		fname := filepath.Base(key)
-
-		wr, err := opts.ReportBucket.NewWriter(ctx, fname, nil)
-
-		if err != nil {
-			return fmt.Errorf("Failed to create new writer for processing report, %w", err)
-		}
-
-		_, err = wr.Write(rsp_body)
-
-		if err != nil {
-			return fmt.Errorf("Failed to write processing report, %w", err)
-		}
-
-		err = wr.Close()
-
-		if err != nil {
-			return fmt.Errorf("Failed to close processing report after writing, %w", err)
-		}
-
-		return nil
-	}
-
-	cfg := opts.Config
-
-	dest_cache, err := iiifcache.NewDerivativesCacheFromConfig(cfg)
-
-	if err != nil {
-		return fmt.Errorf("Failed to derive derivatives cache for processing report, %w", err)
-
-	}
-
-	err = dest_cache.Set(key, rsp_body)
-
-	if err != nil {
-		return fmt.Errorf("Failed to write report, %w", err)
-	}
-
-	slog.Debug("Wrote processing report file", "path", key)
-
-	// START OF HTML version
-
-	if opts.GenerateReportHTML {
-
-		report_html, err := iiifprocess.GenerateProcessReportHTML(ctx, rsp_body)
-
-		html_root := filepath.Dir(key)
-		html_path := filepath.Join(html_root, "index.html")
-
-		err = dest_cache.Set(html_path, report_html)
-
-		if err != nil {
-			return fmt.Errorf("Failed to write HTML %s, %w", html_path, err)
-		}
 	}
 
 	return nil
